move parallel capabilities to download_images
