+++ examples/nips17_adversarial_competition/dataset/README.md
@@ -56,12 +56,8 @@ OUTPUT_DIR=images
 # Download images
 python download_images.py --input_file=${CSV_FILE} --output_dir=${OUTPUT_DIR}
 ```
-
-To accelerate the image downloading, `download_images_multiprocessed.py` spawns multiple processes.
-It can be run using the same parameters as the sequential version:
-```
-python download_images_multiprocessed.py --input_file=${CSV_FILE} --output_dir=${OUTPUT_DIR}
-```
+To accelerate the image downloading, multiple threads are spawned (the default is CPU_COUNT + 1).
+You can change that setting `--threads` to an other positive value.
 
 All downloaded images will be cropped according to the bounding boxes in
 `dev_dataset.csv` and resized to 299x299 pixels.
